Decouple language dropdown from navbar and footer additions
